Fixed some potential bugs.
Removed a hasAccount check from add, subtract and hasEnough. These checks should not exist.
The reasons are:
hasEnough: The default for an account may not have enough to pay for anything. Changed to check if Server Account instead.
add & subtract: Changed the check to only isEmpty. This is because money should not just be created or destroyed.