Make chestshop buying transactions fail if money deposit on seller's account fails.
Some economy plugins, such as Gringotts, only allow a limited amount of money in an account.
Thus, deposits to an account can fail. Chestshop would ignore the failure, which could cause a seller not
to receive the payment for selling an item.

This commit fixes the problem by canceling a transaction if the seller is not able to receive the money of a sale.